Extract symbol usage analysis from exception analyzer
This creates an extension method and context to register actions that
will be invoked for usages of symbols (the built-in RegisterSymbol
action only deals with defined symbols).
